Create server.py
Server program has all the logic to control and regulate the Chat, so most of the chat logic is implemented with a server program. So first step of communication is to identify the users, how server do this? In network communication, users are identified by a socket which is nothing but a combination of IP address and port address. So, for human understanding, Mike and Lisa will be chatting but for a network, it is two sockets process which is sending and receiving bytes.